estime only on selected boat (a toggle is available to fore estime drawing for unselected boat) zoom to grib on load is now parameter, default = no zoom some cleaning in projection position of buttons in VLM2 updated to better differenciate prtho pilot and pilototo bug fix in grib in function to automatically select the right date when current date is out of the grib switch to double instead of float for some data
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@19 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/GribPlot.cpp
@@ -357,7 +357,7 @@ void GribPlot::draw_RAIN_Color(QPainter &pnt, const Projection *proj, bool smoot
         
     
     int i, j;
-    float x, y, v;
+    double x, y, v;
     int W = proj->getW();
     int H = proj->getH();
 
@@ -409,7 +409,7 @@ void GribPlot::draw_CLOUD_Color(QPainter &pnt, const Projection *proj, bool smoo
         return;
     
     int i, j;
-    float x, y, v;
+    double x, y, v;
     int W = proj->getW();
     int H = proj->getH();
     QRgb   rgb;
@@ -444,7 +444,7 @@ void GribPlot::draw_HUMID_Color(QPainter &pnt, const Projection *proj, bool smoo
         return;
     
     int i, j;
-    float x, y, v;
+    double x, y, v;
     int W = proj->getW();
     int H = proj->getH();
     QRgb   rgb;
@@ -484,7 +484,7 @@ void GribPlot::draw_WIND_Color(QPainter &pnt, const Projection *proj, bool smoot
         
     
     int i, j;
-    float x, y, vx, vy, v;
+    double x, y, vx, vy, v;
     int W = proj->getW();
     int H = proj->getH();
     
@@ -532,7 +532,7 @@ void GribPlot::draw_WIND_Arrows
     if (recx == NULL || recy == NULL)
         return;        
     int i, j;
-    float x, y, vx, vy;
+    double x, y, vx, vy;
     int W = proj->getW();
     int H = proj->getH();
     int space =  barbules ? windBarbuleSpace : windArrowSpace;
@@ -860,7 +860,7 @@ void GribPlot::draw_TEMPERATURE_Labels(QPainter &pnt, const Projection *proj)
     pnt.setFont(fontTemperatureLabels);
     pnt.setPen(QColor(0,0,0));
     
-    float x, y, v;
+    double x, y, v;
     int i, j, dimin, djmin;
     dimin = 50;
     djmin = 30;
